requirements.txt
arabic-reshaper==2.1.4
asgiref==3.4.1
asn1crypto==1.5.1
attrs==22.1.0
backports.zoneinfo==0.2.1
Brotli==1.0.9
certifi==2022.12.7
cffi==1.15.0
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.4
cssselect2==0.5.0
Django==4.0.3
django-admin-interface==0.22.2
django-colorfield==0.7.2
django-debug-toolbar==3.8.1
django-flat-responsive==2.0
django-flat-theme==1.1.4
fonttools==4.31.2
future==0.18.2
gunicorn==20.1.0
html5lib==1.1
idna==3.4
iniconfig==1.1.1
lxml==4.9.1
mysqlclient==2.1.0
oscrypto==1.3.0
packaging==21.3
Pillow==9.0.1
pluggy==1.0.0
psycopg2==2.9.3
py==1.11.0
pycparser==2.21
pydyf==0.1.2
pyHanko==0.15.1
pyhanko-certvalidator==0.19.6
pyparsing==3.0.9
PyPDF3==1.0.6
pyphen==0.12.0
pytest==6.2.5
pytest-django==4.5.2
python-bidi==0.4.2
python-decouple==3.6
pytz==2021.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
reportlab==3.6.12
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
svglib==1.4.1
tinycss2==1.1.1
toml==0.10.2
tqdm==4.64.1
tzdata==2022.7
tzlocal==4.2
uritools==4.0.0
urllib3==1.26.13
weasyprint==54.2
webencodings==0.5.1
whitenoise==6.0.0
xhtml2pdf==0.2.8
